front-end/requirements.txt
google-cloud-storage==2.9.0
google-cloud-aiplatform==1.43.0
google-cloud-vision
streamlit==1.31.1
st-pages==0.4.5
